backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "catchmyvibe"
version = "0.1.0"
description = "DJ Enhancement Platform - Aggregate, Analyze, Recommend"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "CatchMyVibe Team" }
]
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # Task Queue
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",

    # Audio Processing
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "mutagen>=1.47.0",
    "numpy>=1.26.3",

    # Storage
    "google-api-python-client>=2.115.0",
    "google-auth-oauthlib>=1.2.0",
    "boto3>=1.34.25",
    "httpx>=0.26.0",

    # Streaming APIs
    "spotipy>=2.23.0",
    "tidalapi>=0.7.3",

    # ML & Embeddings
    "scikit-learn>=1.4.0",
    "torch>=2.1.2",
    "transformers>=4.37.2",

    # Utilities
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.3",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.13",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
catchmyvibe = "api.main:run"

[tool.hatch.build.targets.wheel]
packages = ["api", "ingest", "analysis", "recommend", "models", "storage"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["api", "ingest", "analysis", "recommend", "models", "storage"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
